parse cmd line args, uint32 max accepted for id team

+++ src/handle_board.c
@@ -5,7 +5,7 @@
  * @param vector
  * @return index
 */
-u_int16_t get_board_index (t_vec vec)
+uint32_t get_board_index (t_vec vec)
 {
 	int idx = (vec.y * BOARD_W) +  vec.x;
 	if (idx < 0 || idx >= BOARD_SIZE) {
@@ -21,7 +21,7 @@ u_int16_t get_board_index (t_vec vec)
  * @param position vector
  * @param value to set
 */
-void set_tile_board_val(u_int16_t *array, t_vec vec, u_int16_t value)
+void set_tile_board_val(uint32_t *array, t_vec vec, uint32_t value)
 {
 	int idx = get_board_index(vec);
 	if (idx == OUT_OF_BOARD) {
@@ -31,14 +31,14 @@ void set_tile_board_val(u_int16_t *array, t_vec vec, u_int16_t value)
 }
 
 /* @brief display board in stdout */
-void display_uint16_array(u_int16_t *array)
+void display_uint16_array(uint32_t *array)
 {
 	int i = 0, j = 0;
 
 	for (i = 0; i < BOARD_H; i++) {
 		for (j = 0; j < BOARD_W; j++) {
 			int idx = i * BOARD_W + j;
-			ft_printf_fd(1, YELLOW"[%s%s%d%s%s] "RESET, RESET, CYAN, array[idx], RESET, YELLOW);
+			ft_printf_fd(1, YELLOW"[%s%s%u%s%s] "RESET, RESET, CYAN, array[idx], RESET, YELLOW);
 		}
 		ft_printf_fd(1, "\n");
 	}
